Made proper changes

+++ lua/custom/plugins/code_runner.lua
@@ -9,7 +9,7 @@ return {
     local remove_command = 'rm'
     local file_to_remove = ' /tmp/$fileNameWithoutExt'
 
-    if vim.fn.has 'win32' then
+    if vim.fn.has 'win32' and not vim.fn.has 'mac' then
       start_command = 'g++'
       remove_command = 'del /F'
       file_to_remove = ' \\tmp\\$fileNameWithoutExt.exe'
